Add from_hexstring() to blob_t and fix old model tests
