+++ irohad/model/converters/impl/pb_query_factory.cpp
@@ -136,8 +136,8 @@ namespace iroha {
         const auto &pb_sign = pb_query.signature();
 
         Signature sign{};
-        sign.pubkey = pubkey_t::from_string(pb_sign.public_key());
-        sign.signature = sig_t::from_string(pb_sign.signature());
+        sign.pubkey = pubkey_t::from_hexstring(pb_sign.public_key());
+        sign.signature = sig_t::from_hexstring(pb_sign.signature());
 
         val->query_counter = pl.meta().query_counter();
         val->signature = sign;
