+++ irohad/model/converters/impl/pb_transaction_factory.cpp
@@ -47,8 +47,8 @@ namespace iroha {
 
         for (const auto &pb_sig : pb_tx.signatures()) {
           model::Signature sig{};
-          sig.pubkey = pubkey_t::from_string(pb_sig.public_key());
-          sig.signature = sig_t::from_string(pb_sig.signature());
+          sig.pubkey = pubkey_t::from_hexstring(pb_sig.public_key());
+          sig.signature = sig_t::from_hexstring(pb_sig.signature());
           tx.signatures.push_back(sig);
         }
